requirements.txt
streamlit
pandas
matplotlib
fpdf2
streamlit-lottie
requests